Added drag and drop file upload feature
The user interface has been improved with the addition of a drag and drop zone for file uploads. To enhance user experience, an overlay is displayed to guide the user where to drop files. Additionally, the user interaction was upgraded by updating and handling drag, dragleave and drop events in javascript